Add filter pushdown tests and extract widening tests
Add filter pushdown tests for type widening scenarios covering
INT->Integer, INT->DOUBLE, INT->Decimal, Decimal->Decimal, and
INT32 narrowing with various filter types (BigintRange, DoubleRange,
IsNull, IsNotNull). 6 tests DISABLED due to known filter pushdown
limitations (DoubleRange/HugeintRange crash, rescale mismatch).

Extract all widening/narrowing/filter tests and FloatToDoubleEvolutionTest
from ParquetReaderTest.cpp to ParquetReaderWideningTest.cpp.
